requirements.txt
openai
langchain
langchain-openai
langchain-core
langchain-community
langchain-groq
langchainhub
streamlit
python-dotenv
requests
pypdf
bs4
wikipedia
langchain-anthropic
arxiv
duckduckgo-search
langchain-chroma
faiss-cpu
# chromadb
langchain-ollama
langserve
fastapi
uvicorn
sse_starlette
pdf2image
layoutparser
streamlit-lottie
google-search-results
langchain-tavily
langchain_experimental
tavily-python
pymupdf
rank_bm25
pydantic_ai

langgraph

Ipython
PyPDF2